Upgraded Serial Plotting and Transmitting
Using a union, I created a special float type that was byte addressable.
This way, I am able to easily and clearly transmit floats over the
serial pport to the PSoC Bridge Control Panel by Cypress. This tool
allows for detailed and adjustable plotting from values over the serial
port, and allows for such features as saving plots to be used in later
analysis. This tool is much better than the Arduino serial plot tool.